@@ -26,11 +26,12 @@
         "phayes/geophp": "~1.2"
     },
     "require-dev": {
-        "mockery/mockery": "^0.9.1",
-        "phpunit/phpunit": "~4.0|~5.0",
+        "mockery/mockery": "^1",
+        "phpunit/phpunit": "~7.0",
         "psy/psysh": "^0.5.1",
-        "satooshi/php-coveralls": "^0.6.1|^1",
-        "symfony/var-dumper": "~2.7|~3.0"
+        "satooshi/php-coveralls": "^1",
+        "symfony/thanks": "^1.0",
+        "symfony/var-dumper": "~3.0"
     },
     "autoload": {
         "psr-4": {
