@@ -5,25 +5,18 @@ namespace Spinen\Geometry;
 use ArrayAccess;
 use Countable;
 use Iterator;
-use Mockery;
+use Mockery\Adapter\Phpunit\MockeryPHPUnitIntegration;
 use Mockery\MockInterface;
-use PHPUnit_Framework_TestCase;
+use PHPUnit\Framework\TestCase as PHPUnitTestCase;
 
 /**
  * Class TestCase
  *
  * @package Tests\Spinen\BrowserFilter
  */
-abstract class TestCase extends PHPUnit_Framework_TestCase
+abstract class TestCase extends PHPUnitTestCase
 {
-    public function tearDown()
-    {
-        if (class_exists('Mockery')) {
-            Mockery::close();
-        }
-
-        parent::tearDown();
-    }
+    use MockeryPHPUnitIntegration;
 
     /**
      * Helper to allow mocking of iterator classes.
@@ -31,7 +24,7 @@ abstract class TestCase extends PHPUnit_Framework_TestCase
      * @link https://gist.github.com/VladaHejda/8299871
      *
      * @param MockInterface $mock
-     * @param array         $items
+     * @param array $items
      *
      * @return void
      */
@@ -80,7 +73,7 @@ abstract class TestCase extends PHPUnit_Framework_TestCase
 
             $mock->shouldReceive('next')
                  ->andReturnUsing(function () use (& $counter) {
-                     ++ $counter;
+                     ++$counter;
                  });
         }
 
